refactor: extract IP-comparison and CDN-pulling renderers
